Anlage Api Projekt und Refactoring
