+++ DataInputt/DataInputt.ZeitService.Api/Time.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace DataInputt.ZeitService.Api
+{
+    public class Time
+    {
+        public DateTime Start { get; set; }
+        public DateTime End { get; set; }
+        public string Project { get; set; }
+        public int uId { get; set; }
+        public int Id { get; set; }
+    }
+}
