+++ DataInputt/DataInputt.ZeitService.Api/User.cs
@@ -1,0 +1,8 @@
+﻿namespace DataInputt.ZeitService.Api
+{
+    public class User
+    {
+        public string Name { get; set; }
+        public string Passwort { get; set; }
+    }
+}
