+++ DataInputt/DataInputt.sln
@@ -14,6 +14,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WcfServer", "WcfServer\WcfS
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApi", "WebApi\WebApi.csproj", "{B81127E7-3371-4180-8555-74D4F9F95334}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataInputt.ZeitService.Api", "DataInputt.ZeitService.Api\DataInputt.ZeitService.Api.csproj", "{9DB08342-AB29-485D-AE9F-0D980C2C43B7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -36,6 +38,10 @@ Global
 		{B81127E7-3371-4180-8555-74D4F9F95334}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B81127E7-3371-4180-8555-74D4F9F95334}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B81127E7-3371-4180-8555-74D4F9F95334}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9DB08342-AB29-485D-AE9F-0D980C2C43B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9DB08342-AB29-485D-AE9F-0D980C2C43B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9DB08342-AB29-485D-AE9F-0D980C2C43B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9DB08342-AB29-485D-AE9F-0D980C2C43B7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
