+++ DataInputt/DataInputt/TimeRepo.cs
@@ -1,5 +1,6 @@
 ﻿using DataInputt.Models;
 using DataInputt.ZeitService;
+using DataInputt.ZeitService.Api;
 using System;
 using System.Collections.Generic;
 using System.Linq;
